add edit profile and more funtionalidy

+++ controllers/videoController.js
@@ -1,8 +1,7 @@
 import routes from '../routes';
-import Video from '../models/Videos';
+import Video from '../models/Video';
 
 // Home
-
 export const home = async (req, res) => {
    // Find all the videos in our DB
    try {
@@ -15,7 +14,6 @@ export const home = async (req, res) => {
 };
 
 // Search
-
 export const search = async (req, res) => {
    // Identify the word in the search
    const {
@@ -34,7 +32,6 @@ export const search = async (req, res) => {
 };
 
 // Upload
-
 export const getUpload = (req, res) =>
    res.render('upload', { pageTitle: 'Upload' });
 
@@ -47,19 +44,20 @@ export const postUpload = async (req, res) => {
       fileUrl: path,
       title,
       description,
+      creator: req.user.id,
    });
+   req.user.videos.push(newVideo.id);
+   req.user.save();
    res.redirect(routes.videoDetail(newVideo.id));
 };
 
 // Video Detail
-
 export const videoDetail = async (req, res) => {
    const {
       params: { id },
    } = req;
-
    try {
-      const video = await Video.findById(id);
+      const video = await Video.findById(id).populate('creator');
       res.render('videoDetail', { pageTitle: video.title, video });
    } catch (error) {
       res.redirect(routes.home);
@@ -67,15 +65,17 @@ export const videoDetail = async (req, res) => {
 };
 
 // Editing a video
-
 export const getEditVideo = async (req, res) => {
    const {
       params: { id },
    } = req;
-
    try {
       const video = await Video.findById(id);
-      res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+      if (String(video.creator) !== req.user.id) {
+         throw Error();
+      } else {
+         res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+      }
    } catch (error) {
       res.redirect(routes.home);
    }
@@ -86,7 +86,6 @@ export const postEditVideo = async (req, res) => {
       params: { id },
       body: { title, description },
    } = req;
-
    try {
       await Video.findOneAndUpdate({ _id: id }, { title, description });
       res.redirect(routes.videoDetail(id));
@@ -96,16 +95,19 @@ export const postEditVideo = async (req, res) => {
 };
 
 // Delete a video
-
 export const deleteVideo = async (req, res) => {
    const {
       params: { id },
    } = req;
-
    try {
-      await Video.findOneAndRemove({ _id: id });
+      const video = await Video.findById(id);
+      if (video.creator !== req.user.id) {
+         throw Error();
+      } else {
+         await Video.findOneAndRemove({ _id: id });
+      }
    } catch (error) {
       console.log(error);
+      res.redirect(routes.home);
    }
-   res.redirect(routes.home);
 };
